metirc/segmentation/requirements.txt
torch
torchvision
transformers
pillow
numpy
tqdm
